requirements.txt
# cat requirements.txt
cdflib>=0.3.20
matplotlib>=3.4.1
numba>=0.53.1
numpy>=1.20.2
pandas>=1.2.4
Sphinx>=3.5.4
scipy>=1.6.2
setuptools>=56.0.0
xarray>=0.17.0
python-dateutil>=2.8.1
pyrfu>=2.3.5
geopack>=1.0.8
PyYAML~=6.0
uncertainties~=3.1.6